# Optimize OpenAPI parser performance with single-pass schema processing