FastMCP's OpenAPI parser was experiencing severe performance bottlenecks when processing large schemas like GitHub's 10MB API specification, taking multiple minutes to parse. The primary culprits were expensive deep copy operations in `compress_schema()` and inefficient tree traversals in `_adjust_union_types()`.

This change delivers a 98.6% performance improvement, reducing parse time from multiple minutes to under 2 seconds through systematic optimization.

The main issue was that `compress_schema()` was calling `copy.deepcopy()` on the entire schema before processing, creating massive object copies for schemas that were already 10MB+ in size. Additionally, the function was making three separate passes through the schema tree for different cleanup operations, and `_adjust_union_types()` was recursively traversing deeply nested structures without early termination.

The solution combines three separate tree traversals into a single pass that operates in-place on the original schema. We added early detection to skip union type processing entirely when no `oneOf` keys exist in the schema, and implemented depth limits to prevent exponential slowdown on deeply nested structures. For performance-critical paths, we disabled expensive unused definition removal where schema size reduction provides minimal benefit.

```python
fastmcp.settings.experimental.enable_new_openapi_parser = True

# Before: multiple minutes for large schemas
server = FastMCP.from_openapi(github_schema, client)

# After: ~1.8 seconds for the same schema  
server = FastMCP.from_openapi(github_schema, client)
```

The optimizations maintain full backward compatibility and functionality. A comprehensive performance regression test now ensures GitHub's full API schema parses within 3 seconds, preventing future performance degradation.

Fixes #1195
Inspired by analysis in #1196
Builds on work in #1209 

Here's a test file for the GitHub API spec mentioned in #1195:

```python
import asyncio
import time

import httpx

import fastmcp
from fastmcp import FastMCP

# disable for legacy
fastmcp.settings.experimental.enable_new_openapi_parser = True


async def test():
    # Download the full GitHub API schema (typically ~10MB)
    response = httpx.get(
        "https://raw.githubusercontent.com/github/rest-api-description/refs/heads/main/descriptions-next/ghes-3.17/ghes-3.17.json",
        timeout=30.0,  # Allow time for download
    )
    response.raise_for_status()
    schema = response.json()

    # Time the parsing operation
    start_time = time.time()

    # This should complete quickly with our optimizations
    mcp_server = FastMCP.from_openapi(schema, httpx.AsyncClient())

    elapsed_time = time.time() - start_time

    print(f"OpenAPI parsing took {elapsed_time:.2f}s")

    # Verify the server was created successfully
    assert mcp_server is not None

    # Performance regression test: should complete in under 10 seconds
    assert elapsed_time < 10.0, (
        f"OpenAPI parsing took {elapsed_time:.2f}s, exceeding 10s limit. "
        f"This suggests a performance regression."
    )


if __name__ == "__main__":
    asyncio.run(test())

```